+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- Add `EquipmentUsed` and `TankData` models for dive-specific equipment
+- Add `equipmentused` property to `Dive` struct for gas mix references per dive
+
 ## 0.1.0 - 2026-01-07
 
 ### Added